Added BackendTest and touched up other test files
